requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
requests
beautifulsoup4
pandas
sentence-transformers
faiss-cpu
